Extract AbstractBearerAuthenticator
